leaks not fixed on ft_strjoin

+++ libft/ft_strjoin.c
@@ -21,8 +21,8 @@ char	*ft_strjoin(char const *s1, char const *s2)
 	j = 0;
 	if (!s1 || !s2)
 		return (NULL);
-	i = ft_strlen(s1) + ft_strlen(s2);
-	if (!(str = (char*)malloc(sizeof(char) * (i + 1))))
+	i = (int)(ft_strlen(s1) + ft_strlen(s2));
+	if (!(str = (char*)malloc(sizeof(char) * i + 1)))
 		return (NULL);
 	i = 0;
 	while (s1[i] || s2[j])
